+++ CHANGES.md
@@ -35,8 +35,7 @@
 -   pat calendar: Store the current date and view in query parameters.
 -   pat calendar: Fetch events from the backend.
 -   pat calendar: Allow filtering/hiding events based in comparing the checkbox id with the classes of the displayed events.
--   pat calendar: Support injection of events when clicking on and event rather than redirecting to them.
-    Done by adding `pat-inject` to rendered events via some configuration options.
+-   pat calendar: Support `pat-inject` on events with a URL via `pat-inject-source` and `pat-inject-target` configuration options.
 -   pat calendar: Support `pat-switch` for rendered events via some configuration options.
 -   pat calendar: Store view, date and active categories per URL, allowing to individually customize the calendar per page.
 -   pat tooltip: Use tippy v6 based implementation.